% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_das_df.R
\name{as_das_df}
\alias{as_das_df}
\alias{as_das_df.das_df}
\alias{as_das_df.data.frame}
\title{Coerce object to a das_df object}
\usage{
as_das_df(x)

\method{as_das_df}{das_df}(x)

\method{as_das_df}{data.frame}(x)
}
\arguments{
\item{x}{an object to be coerced to class \code{das_df}}
}
\value{
An object of class \code{das_df}
}
\description{
Check if an object is of class \code{das_df}, or coerce it if possible.
}
\details{
Only data frames can be coerced to an object of class \code{das_df}.
If \code{x} does not have column names and classes as specified in \code{\link{das_df-class}},
then the function returns an error message detailing the first column that does not
meet the requirements of a \code{das_df} object.
}
\seealso{
\code{\link{das_df-class}}
}
