% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_spcodes.R
\name{das_spcodes}
\alias{das_spcodes}
\alias{das_spcodes_read}
\title{Read DAS SpCodes file}
\usage{
das_spcodes_read(file, skip = 0)
}
\arguments{
\item{file}{character; filename of .dat file from which to read
species codes}

\item{skip}{integer; default is 3.
Number of lines to skip when reading SpCodes file.
See \code{\link[readr:read_fwf]{readr::read_fwf()}} for more details}
}
\value{
Data frame with four columns:
\itemize{
\item SpCode: species code, columns 1 to 4
\item Abbr: species abbreviation, columns 6 to 15
\item SciName: species scientific name, columns 18 to 57
\item CommonName: species common name, columns 58 until the end of the line
}
}
\description{
Read DAS SpCodes file
}
\details{
Provide a standardized function to read a shipboard DAS SpCodes file.
Methods described in 'returns'
}
\examples{
sp.codes.file <- system.file("extdata", "SpCodes_sample.dat", package = "swfscDAS")
das_spcodes_read(sp.codes.file)

}
