% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_within_strata.R
\name{das_within_strata}
\alias{das_within_strata}
\alias{das_within_strata.list}
\alias{das_within_strata.data.frame}
\title{DAS strata - points}
\usage{
das_within_strata(x, ...)

\method{das_within_strata}{list}(x, y, ...)

\method{das_within_strata}{data.frame}(x, y, x.lon = "Lon", x.lat = "Lat", ...)
}
\arguments{
\item{x}{a list or a data frame (including an object of class \code{das_df}).
If \code{x} is a list, then it must be the output of
\code{\link{das_effort}} or \code{\link{das_effort_sight}}.
If \code{x} is a data frame, the user must also specify
the columns of \code{x} with the relevant coordinates
using \code{x.lon} and \code{x.lat}.}

\item{...}{ignored}

\item{y}{list of file path(s) of the CSV(s) that contain points defining the strata.
The list may be named; see 'Value' section for how these names are used}

\item{x.lon}{character; name of the longitude column of \code{x}.
Ignored if \code{x} is a list; default is "Lon"}

\item{x.lat}{character; name of the latitude column of \code{x}.
Ignored if \code{x} is a list; default is "Lat"}
}
\value{
Columns are added to \code{x} indicating if each point was
  in (\code{TRUE}) or out (\code{FALSE}) of the corresponding strata polygon.
  The names of these columns are the names of \code{y};
  the element(s) of \code{y} will have the name InPoly#,
  where '#' is the index of that strata polygon in \code{y}.

  If \code{x} is a list (i.e. the output of one of the effort functions),
  then the strata columns are added to both the segdata and sightinfo data frames.
  However, note that the columns added to the sightinfo data frame still indicate
  whether or not the corresponding segment midpoint was in the strata,
  not the sighting point itself.
}
\description{
Determine the strata for swfscDAS outputs
}
\details{
Assigns DAS event points or segment midpoints to strata polygon.
  If \code{x} is a list, then 1) it must be the output of
  \code{link{das_effort}} or \code{link{das_effort_sight}} and
  2) the segment midpoints (column names mlon and mlat, respectively)
  are the points checked if they are in the strata.
  If \code{x} is a data frame, then the user must provide the columns
  that specify the point coordinates to check.

  \code{x} should not be an object of class \code{das_dfr},
  or an object of class \code{das_df} created with \code{add.dtll.sight = FALSE},
  because the ? and numeric event codes will have NA latitude and longitude values.
}
\examples{
y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)
y.eff <- das_effort(y.proc, method = "section", num.cores = 1)

strata.file <- system.file("das_sample_strata.csv", package = "swfscDAS")
das_within_strata(y.eff, list(InPoly = strata.file), x.lon = "Lon", x.lat = "Lat")

das_within_strata(y.proc, list(strata.file))

\dontrun{
# Visualize effort midpoints and strata polygon
require(sf)
y.eff.strata <- das_within_strata(y.eff, list(InPoly = strata.file))
segdata <- st_as_sf(y.eff.strata$segdata, coords = c("mlon", "mlat"), crs = 4326)

# Make strata polygon
strata.df <- read.csv(strata.file)
strata.sfc <- st_sfc(
  st_polygon(list(matrix(c(strata.df$Lon, strata.df$Lat), ncol = 2))),
  crs = 4326
)

plot(segdata["InPoly"], axes = TRUE, reset = FALSE,
     xlim = c(-137, -142.5), ylim = c(42, 47))
plot(strata.sfc, add = TRUE)
}

}
