% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_greatcircle.R
\name{distance_greatcircle}
\alias{distance_greatcircle}
\title{Calculate great-circle distance}
\usage{
distance_greatcircle(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{numeric; starting latitude coordinate(s)}

\item{lon1}{numeric; starting longitude coordinate(s)}

\item{lat2}{numeric; ending latitude coordinate(s)}

\item{lon2}{numeric; ending longitude coordinate(s)}
}
\value{
Distance in kilometers between lat1/lon1 and lat2/lon2
}
\description{
Calculate the great-circle distance between two lat/lon points
}
\seealso{
\url{https://en.wikipedia.org/wiki/Great-circle_distance}
}
