% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_chop_condition.R, R/das_chop_equal.R,
%   R/das_effort.R
\name{swfscAirDAS-funcs}
\alias{swfscAirDAS-funcs}
\alias{.chop_condition_eff}
\alias{.chop_equal_eff}
\alias{.dist_from_prev}
\alias{.fn.grcirclkm}
\title{Functions exported only to be used internally by swfscAirDAS}
\usage{
.chop_condition_eff(
  i,
  call.x,
  call.cond.names,
  call.seg.km.min,
  call.func1,
  call.func2
)

.chop_equal_eff(i, call.x, call.seg.km, call.r.pos, call.func1, call.func2)

.dist_from_prev(z, z.dist.method)

.fn.grcirclkm(lat1, lon1, lat2, lon2)
}
\arguments{
\item{i}{ignore}

\item{call.x}{ignore}

\item{call.cond.names}{ignore}

\item{call.seg.km.min}{ignore}

\item{call.func1}{ignore}

\item{call.func2}{ignore}

\item{call.seg.km}{ignore}

\item{call.r.pos}{ignore}

\item{z}{ignore}

\item{z.dist.method}{ignore}

\item{lat1}{ignore}

\item{lon1}{ignore}

\item{lat2}{ignore}

\item{lon2}{ignore}
}
\description{
Functions exported only to be used internally by swfscAirDAS
}
