% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_das_df.R
\name{as_das_df}
\alias{as_das_df}
\alias{as_das_df.das_df}
\alias{as_das_df.data.frame}
\title{Convert object to a das_df object}
\usage{
as_das_df(x)

\method{as_das_df}{das_df}(x)

\method{as_das_df}{data.frame}(x)
}
\arguments{
\item{x}{object to be converted to a `das_df` object}
}
\value{
an object of class `das_df`
}
\description{
Convert object to a das_df object
}
\details{
todo
}
