% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_das_dfr.R
\name{as_das_dfr}
\alias{as_das_dfr}
\alias{as_das_dfr.das_dfr}
\alias{as_das_dfr.data.frame}
\title{Coerce object to a das_dfr object}
\usage{
as_das_dfr(x)

\method{as_das_dfr}{das_dfr}(x)

\method{as_das_dfr}{data.frame}(x)
}
\arguments{
\item{x}{an object to be coerced to class \code{das_dfr}}
}
\value{
An object of class `das_dfr`
}
\description{
Check if an object is of class \code{das_dfr}, or coerce it if possible.
}
\details{
Only data frames can be coerced to an object of class \code{das_dfr}.
  If \code{x} does not have column names and classes as specified in \code{\link{das_dfr-class}},
  then the function returns an error message detailing the first column that does not
  meet the requirements of a \code{das_dfr} object.
}
\seealso{
\code{\link{das_dfr-class}}
}
