% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_das_dfr.R
\name{as_das_dfr}
\alias{as_das_dfr}
\alias{as_das_dfr.das_dfr}
\alias{as_das_dfr.data.frame}
\title{Convert object to a das_dfr object}
\usage{
as_das_dfr(x)

\method{as_das_dfr}{das_dfr}(x)

\method{as_das_dfr}{data.frame}(x)
}
\arguments{
\item{x}{object to be converted to a `das_dfr` object}
}
\value{
an object of class `das_dfr`
}
\description{
Convert object to a das_dfr object
}
\details{
todo
}
