% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_check.R
\name{das_check}
\alias{das_check}
\title{Check DAS file}
\usage{
das_check(file, file.out = NULL)
}
\arguments{
\item{file}{filename(s) of one or more DAS files}

\item{file.out}{filename to which to write the error log; default is NULL}
}
\value{
A data frame with three columns: the line number,
'ID' (columns 4-39 form the DAS file), and description of the issue

If \code{file.out} is not \code{NULL}, then the error log is also
written to a text file
}
\description{
Check that DAS file has accepted values
}
\details{
Precursor to DASCHECK. Checks that the following is true:
\itemize{
  \item Event codes are one of the following: #, *, ?, 1, 2, 3, 4, 5, 6, 7, 8,
    A, B, C, E, F, k, K, N, P, Q, R, s, S, t, V, W, g, p, X, Y, Z.
  \item The effort dot matches effort determined using B, R, and E events
  \item And event/column pairs meet the following requirements:
}

\tabular{llll}{
  \emph{Item} \tab \emph{Event} \tab \emph{Column} \tab \emph{Requirement}\cr
  Cruise number  \tab B \tab Data1 \tab Can be converted to a numeric value\cr
  Mode           \tab B \tab Data2 \tab Must be one of C, P, c, p, or NA (blank)\cr
  Effort type    \tab R \tab Data1 \tab Must be one of F, N, S, or NA (blank)\cr
  Course         \tab N \tab Data1 \tab Can be converted to a numeric value\cr
  Beaufort       \tab V \tab Data1 \tab Can be converted to a numeric value\cr
  Swell height   \tab V \tab Data2 \tab Can be converted to a numeric value\cr
  Rain or fog    \tab W \tab Data1 \tab Can be converted to a numeric value\cr
  Horizontal sun \tab W \tab Data2 \tab Can be converted to a numeric value\cr
  Vertical sun   \tab W \tab Data3 \tab Can be converted to a numeric value\cr
  Visibility     \tab W \tab Data5 \tab Can be converted to a numeric value\cr
  NA \tab S, K, M \tab Data3-7    \tab Can be converted to a numeric value \cr
  NA \tab s, k    \tab Data2-5    \tab Can be converted to a numeric value \cr
  NA \tab t       \tab Data3-5, 7 \tab Can be converted to a numeric value \cr
  NA \tab F       \tab Data2-4    \tab Can be converted to a numeric value \cr
  NA \tab 1-8     \tab Data2-8    \tab Can be converted to a numeric value \cr
}
}
\examples{
y <- system.file("das_sample.das", package = "swfscDAS")
das_check(y)

}
