% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_check.R
\name{das_check}
\alias{das_check}
\title{Check DAS file}
\usage{
das_check(file, file.out = NULL)
}
\arguments{
\item{file}{filename(s) of one or more DAS files}

\item{file.out}{filename to which to write the error log;
default is \code{NULL}}
}
\value{
A data frame with five columns: the file name, line number,
index (row number) from the \code{das_read(file)} data frame,
'ID' (columns 4-39 from the DAS file), and description of the issue

If \code{file.out} is not \code{NULL}, then the error log is also
written to a text file
}
\description{
Check that DAS file has accepted values
}
\details{
TODO: Add species code and observer length checks for sightings, as in swfscAirDAS

Precursor to DASCHECK. Checks that the following is true:
\itemize{
  \item Event codes are one of the following: #, *, ?, 1, 2, 3, 4, 5, 6, 7, 8,
    A, B, C, E, F, k, K, N, P, Q, R, s, S, t, V, W, G, g, p, X, Y, Z.
  \item The effort dot matches effort determined using B, R, and E events
  \item And event/column pairs meet the following requirements:
}

\tabular{llll}{
  \emph{Item} \tab \emph{Event} \tab \emph{Column} \tab \emph{Requirement}\cr
  Cruise number  \tab B \tab Data1 \tab Can be converted to a numeric value\cr
  Mode           \tab B \tab Data2 \tab Must be one of C, P, c, p, or NA (blank)\cr
  Effort type    \tab R \tab Data1 \tab Must be one of F, N, S, or NA (blank)\cr
  Effective strip width sides \tab R \tab Data2 \tab Must be one of F, H, or NA (blank)\cr
  Course         \tab N \tab Data1 \tab Can be converted to a numeric value\cr
  Beaufort       \tab V \tab Data1 \tab Can be converted to a numeric value\cr
  Swell height   \tab V \tab Data2 \tab Can be converted to a numeric value\cr
  Rain or fog    \tab W \tab Data1 \tab Can be converted to a numeric value\cr
  Horizontal sun \tab W \tab Data2 \tab Can be converted to a numeric value\cr
  Vertical sun   \tab W \tab Data3 \tab Can be converted to a numeric value\cr
  Visibility     \tab W \tab Data5 \tab Can be converted to a numeric value\cr
  Sighting (mammal) \tab S, K, M \tab Data3-7    \tab Can be converted to a numeric value\cr
  Photos \tab A \tab Data3 \tab Must be one of N, Y, n, y, or NA (blank)\cr
  Birds  \tab A \tab Data4 \tab Must be one of N, Y, n, y, or NA (blank)\cr
  Resighting      \tab s, k    \tab Data2-5    \tab Can be converted to a numeric value\cr
  Turtle sighting \tab t       \tab Data3-5, 7 \tab Can be converted to a numeric value\cr
  JFR             \tab t \tab Data6 \tab Must be one of F, J, N, R, or NA (blank)\cr
  Fishing vessel  \tab F       \tab Data2-4    \tab Can be converted to a numeric value\cr
  Sighting info \tab 1-8     \tab Data2-8    \tab Can be converted to a numeric value\cr
  Sighting info \tab 1-8 \tab Data9 \tab The Data9 column must be NA (blank) for events 1-8\cr
}

Outstanding questions:
\itemize{
  \item Documentation says Data8 and Data9 for SKM events be numeric, but currently ~5000 lines are not
  \item What to add?
  \item How to check that everything is right-justified?
}
}
\examples{
y <- system.file("das_sample.das", package = "swfscDAS")
das_check(y)

}
