% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_check.R
\name{das_check}
\alias{das_check}
\title{Check DAS file}
\usage{
das_check(file, file.out = NULL)
}
\arguments{
\item{file}{filename(s) of one or more DAS files}

\item{file.out}{filename to which to write the error log; default is NULL}
}
\value{
A list...
  If file.out is not NULL, then the error log is also written to a text file
}
\description{
Check that DAS file has accepted values
}
\details{
Precursor to DASCHECK. Checks that the following is true:

  \itemize{
    \item Event codes are one of the following: #, *, ?, 1, 2, 3, 4, 5, 6, 7, 8,
      A, B, C, E, F, k, K, N, P, Q, R, s, S, t, V, W, g, p, X, Y, Z.
    \item
  }
  \tabular{lrr}{
    \emph{Item}  \tab \emph{Columns} \tab \emph{Format}\cr
    Event number \tab 1-3\cr
  }
}
\examples{
# file <- system.file("das_sample.das", package = "swfscDAS")
y <- system.file("das_sample.das", package = "swfscDAS")
das_check(y)

das_check("../DAS/AllDas.das")

}
