% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_chop_condition.R
\name{das_chop_condition}
\alias{das_chop_condition}
\alias{das_chop_condition.data.frame}
\alias{das_chop_condition.das_df}
\title{Chop DAS data - condition}
\usage{
das_chop_condition(x, ...)

\method{das_chop_condition}{data.frame}(x, ...)

\method{das_chop_condition}{das_df}(x, seg.km.min = 0.1, dist.method = NULL, ...)
}
\arguments{
\item{x}{\code{das_df} object,
or a data frame that can be coerced to a \code{das_df} object.
This data must be filtered for 'OnEffort' events;
see the Details section below}

\item{...}{ignored}

\item{seg.km.min}{numeric; minimum allowable segment length (in kilometers).
Default is 0.1. See the Details section below for more information}

\item{dist.method}{character;
method to use to calculate distance between lat/lon coordinates.
Can be "greatcircle" to use the great circle distance method,
or one of "lawofcosines", "haversine", or "vincenty" to use
\code{\link[swfscMisc]{distance}}.
Default is \code{NULL} since these distances should have already been
calculated in \code{\link{das_effort}}}
}
\value{
List of two data frames:
\itemize{
  \item \code{x}, with columns added for the corresponding unique segment code and number
  \item segdata: data frame with one row for each segment, and columns with
    relevant data (see \code{\link{das_effort}} for specifics)
}
}
\description{
Chop DAS data into a new effort segment every time a condition changes
}
\details{
This function is intended to only be called by \code{\link{das_effort}}
  when the "condition" method is specified.
  Thus, \code{x} must be filtered for events (rows) where either
  the 'OnEffort' column is \code{TRUE} or the 'Event' column is "E";
  see \code{\link{das_effort}} for more details.
  This function chops each continuous effort section (henceforth 'effort sections')
  in \code{x} into modeling segments (henceforth 'segments') by
  creating a new segment every time a condition changes.
  Each effort section runs from a B/R event to its corresponding E event.
  After chopping, \code{\link{das_segdata_avg}} is called to get relevant
  segdata information for each segment.

  Changes in the following conditions trigger a new segment:
  Beaufort, swell height, rain/fog/haze code, horizontal sun, vertical sun,
  and visibility (no glare because glare is dependent on sun positions).
  The main exception is when multiple condition changes happen at
  the same location, such as a 'BRPVNW' series of events.
  When this happens, no segments of length zero are created;
  rather, a single segment is created that includes all of the condition changes
  (i.e. all of the events in the event series) that happened during
  the series of events (i.e. at the same location).

  In addition, (almost) all segments whose length is less than \code{seg.km.min}
  are combined with the segment immediately following them to ensure that the length
  of (almost) all segments is at least \code{seg.km.min}.
  This allows users to account for situations where multiple conditions,
  such as Beaufort and the visibility, change in rapid succession, say 0.05 km apart.
  When segments are combined, a warning is thrown and the conditions are averaged together
  across the now-larger segment (question - is this right? See below).
  The only exception to this rule is if the short segment ends in an "E" event,
  meaning it is the last segment of the effort section.
  Since in this case there is no 'next' segment, this segment is left as-is.

  Note that the above rule for 'combining' condition changes that have the same location
  into a single segment (such as a 'BRPVNW' series of events)
  is followed even if \code{seg.km.min = 0}.

  If the column \code{dist_from_prev} does not exist
  (it should be calculated and added to \code{x} in \code{\link{das_effort}}),
  then the distance between the lat/lon points of subsequent events
  is calculated using the method specified in \code{dist.method}

  TODO: Make das_segdata_max function so that conditions from tiny segments aren't averaged in
}
\keyword{internal}
