% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_effort.R
\name{das_effort}
\alias{das_effort}
\alias{das_effort.data.frame}
\alias{das_effort.das_df}
\title{Summarize DAS effort}
\usage{
das_effort(x, ...)

\method{das_effort}{data.frame}(x, ...)

\method{das_effort}{das_df}(x, method, sp.codes, dist.method = "vincenty", ...)
}
\arguments{
\item{x}{\code{das_df} object; output from \code{\link{das_process}},
or a data frame that can be coerced to a \code{das_df} object}

\item{...}{arguments passed to the chopping function specified using \code{method}}

\item{method}{character; method to use to chop DAS data into effort segments
Can be "equallength" or "condition" (case-sensitive)}

\item{sp.codes}{character; species code(s) to include in segdata}

\item{dist.method}{character;
method to use to calculate distance between lat/lon coordinates.
Can be "greatcircle" to use the great circle distance method,
or one of "lawofcosines", "haversine", or "vincenty" to use
\code{\link[swfscMisc]{distance}}. Default is "vincenty"}
}
\value{
List of three data frames:
  \itemize{
    \item segdata: one row for every segment, and columns for information including
      unique segment number, start/end/midpoint coordinates, conditions (e.g. Beaufort),
      and number of sightings and number of animals on that segment for every species
      indicated in \code{sp.codes}.
    \item siteinfo: details for all sightings in \code{x}, including:
      the unique segment number it is associated with, segment mid points (lat/lon),
      and whether the sighting was included in the segdata counts (column \code{included}),
      in addition to the other output information described in \code{\link{das_sight}}.
    \item randpicks: see \code{\link{das_chop_equal}}
  }
}
\description{
Chop DAS data into effort segments
}
\details{
This is the top-level function for chopping processed DAS data
  into modeling segments (henceforth 'segments'), and assigning sightings
  and related information (e.g., weather conditions) to each segment.
  This function returns data frames with all relevant information for the
  effort segments and associated sightings ('segdata' and 'siteinfo', respectively).
  Before chopping, the DAS data is filtered for events (rows) where either
  the 'OnEffort' column is \code{TRUE} or the 'Event' column "E".
  In other words, the data is filtered for continuous effort sections (henceforth 'effort sections'),
  where effort sections run from "B"/"R" to "E" events (inclusive),
  and then passed to the chopping function specified using \code{method}.

  TODO
  included: On effort and Beaufort less than or equal to 5
}
\examples{
y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)

das_effort(
  y.proc, method = "equallength", sp.codes = c("016", "018"),
  seg.km = 10, num.cores = 1
)

das_effort(
  y.proc, method = "condition", sp.codes = c("016", "018"),
  seg.km.min = 0.05, num.cores = 1
)

}
