% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_effort_sight.R
\name{das_effort_sight}
\alias{das_effort_sight}
\title{Summarize DAS sightings by effort segment}
\usage{
das_effort_sight(x.list, sp.codes)
}
\arguments{
\item{x.list}{output of \code{\link{das_effort}}; a list with three elements
names 'segdata', 'sightinfo', and 'randpicks', respectively}

\item{sp.codes}{character; species code(s) to include in segdata output.
These must exactly match the species codes in the data,
such as including leading zeros}
}
\value{
A list, identical to \code{x.list} except for
  1) the nSI and ANI columns added to \code{x.list$segdata},
  one each for each element of \code{sp.codes}, and
  2) the included column of \code{x.list$sightinfo}, which has been set as
  \code{FALSE} for sightings of species not listed in \code{sp.codes}
}
\description{
Summarize number of sightings and animals for selected species by segment
}
\details{
This function takes the output of \code{\link{das_effort}} and
  adds columns for the number of sightings (nSI) and number of animals (ANI)
  for selected species (selected via \code{sp.codes}) for each segment
  to the segdata element of \code{x.list}.
  However, only sightings with an included value of \code{TRUE}
  (included is a column in sightinfo) are included in the summaries.
  Having this step separate from \code{\link{das_effort}} allows users to
  personalize the included values as desired for their analysis.

  The ANI columns are the sum of the 'GsSpBest' column output from
  \code{\link{das_sight}}
}
\examples{
y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)
y.cond <- das_effort(
  y.proc, method = "condition",
  conditions = "Bft", seg.min.km = 0.05, num.cores = 1
)

y.cond$sightinfo <- y.cond$sightinfo[y.cond$sightinfo$Event \%in\% c("S", "t"), ]
y.cond$sightinfo$included <- TRUE

das_effort_sight(y.cond, sp.codes = c("013", "076", "DC"))

}
