% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_process.R
\name{das_process}
\alias{das_process}
\alias{das_process.character}
\alias{das_process.data.frame}
\title{Process DAS data}
\usage{
das_process(x, ...)

\method{das_process}{character}(x, ...)

\method{das_process}{data.frame}(x, days.gap = 10, reset.event = TRUE, reset.day = TRUE, ...)
}
\arguments{
\item{x}{either a data frame (the output of \code{\link{das_read}}),
or a character (filepath) which is first passed to \code{\link{das_read}}}

\item{...}{ignored}

\item{days.gap}{numeric of length 1; time gap (in days) used to identify
a new cruise in concatenated DAS files, and thus also when
state/condition information (weather, Bft, Mode, etc) is reset.
Default is 10 days}

\item{reset.event}{logical; indicates if state/condition information
(weather, Bft, Mode, etc) should be reset to \code{NA} if there is an
applicable event with an \code{NA} for that state/condition}

\item{reset.day}{logical; indicates if state/condition information
(weather, Bft, Mode, etc) should be reset to \code{NA} at the beginning of each day.
This argument should only be set to \code{FALSE}
for comparison with older methods, such as Report}
}
\value{
Data frame; the input data frame, i.e. the output of \code{\link{das_read}},
  with the following columns added:
  \tabular{ll}{
    \emph{State/condition}        \tab \emph{Column name}\cr
    On/off effort                 \tab OnEffort\cr
    Cruise number                 \tab Cruise\cr
    Effort mode                   \tab Mode\cr
    Effort type                   \tab EffType\cr
    Course (ship direction)       \tab Course\cr
    Beaufort sea state            \tab Bft\cr
    Swell height (ft)             \tab SwellHght\cr
    Rain/fog/haze code            \tab RainFog\cr
    Horizontal sun (clock system) \tab HorizSun\cr
    Vertical sun (clock system)   \tab VertSun\cr
    Glare                         \tab Glare\cr
    Visibility (nm)               \tab Vis\cr
  }
}
\description{
Process DAS data (the output of \code{\link{das_read}}),
  including extracting state and condition information for each DAS event
}
\details{
If \code{x} is a character, it is assumed to be a filepath and first passed to \code{\link{das_read}}.
  This output is then passed to \code{das_process}.

  DAS data is event-based, meaning most events indicate when a state or weather condition changes.
  For instance, a 'V' event indicates when the Beaufort sea state changes, and
  the Beaufort is the same for subsequent events until the next 'V' event.
  For each state/condition: a new column is created,
  the state/condition information is extracted from relevant events,
  and extracted information is propagated to appropriate subsequent rows (events).
  Thus, each row in the output data frame contains all
  pertinent state/condition information for that row.

  The following assumptions/decisions are made during processing:
  \itemize{
    \item All '#' events (deleted events) are removed
    \item 'Datetime', 'Lat', and 'Lon' information are added to '?' and '1-8' events
    \item Effort is determined as follows: R events turn effort on, and E events turn effort off (TODO: B events??).
      The 'EffortDot' column is ignored
    \item All state/condition information is reset at the beginning of each cruise
      (as identifed using \code{days.gap})
    \item 'Mode' is capitalized, and if 'Mode' is \code{NA}, it is assigned a value of "C"
    \item If 'EffType' is \code{NA}, it is assigned a value of "S"
    \item 'Glare': \code{TRUE} if 'HzSun' is 11, 12 or 1 and 'VtSun' is 2 or 3,
      or if 'HzSun' is 12 and 'VtSun' is 1;
      \code{NA} if 'HzSun' or 'VtSun' is \code{NA}; otherwise \code{FALSE}
    \item Missing values are \code{NA} rather than \code{-1}
  }

  This function was inspired by \code{\link[swfscMisc]{das.read}}
}
\examples{
# TODO

}
\seealso{
For more details about WinCruz, see
  \url{https://swfsc.noaa.gov/uploadedFiles/Divisions/PRD/WinCruz.pdf}
}
