% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_process.R
\name{das_process}
\alias{das_process}
\alias{das_process.character}
\alias{das_process.data.frame}
\title{Process DAS data}
\usage{
das_process(das.df, ...)

\method{das_process}{character}(das.df, ...)

\method{das_process}{data.frame}(das.df, days.gap = 50,
  reset.event = TRUE, reset.day = TRUE, cruise.name = NA,
  cruise.num = NA, ship.name = NA, ...)
}
\arguments{
\item{das.df}{Either a data frame (the output of \link{das_read}) or
a character which is then passed to \link{das_read}}

\item{...}{Ignore}

\item{days.gap}{numeric of length 1; time gap (in days) used to identify
a new cruise in concatenated DAS files. For instance, if days.gap = 50,
rows of das.df will be considered a new cruise if consecutive records are
more than 50 days apart.}

\item{reset.event}{logical; indicates if condition/effort info
(e.g. Bft, Mode, etc) should be reset to NA if there is an
applicable event with an NA for that value}

\item{reset.day}{logical; indicates if condition/effort info
(e.g. Bft, Mode, etc) should be reset to NA at the beginning of each day.
This argument should only be used (i.e. be \code{FALSE}
for comparison with older methods (e.g. Report)}

\item{cruise.name}{character of length 1; name of cruise, e.g. 'MOPS'}

\item{cruise.num}{character or numeric of length 1; cruise number, e.g. 989}

\item{ship.name}{character of length 1; abbr name of ship, e.g. 'DSJ'
TODO: should this be constrained, e.g. must be one of: "DSJ", "END", "MAC", "MACII"?}
}
\value{
data frame with 'carry-over info' columns added (e.g., Mode and Bft)
  OnEffort_RE is On/Off effort as determiend by R and E events: from R event
    to the event before the next E event are considered on effort
  Columns added to data frame:
    CruiseName, Cruise (cruise #), Shipname, OnEffort_RE, Mode, EffType,
    Bft, SwellHght, RainFog, HorizSun, VertSun, Glare, Vis, Course
}
\description{
Process output of \link{das_read}
}
\details{
das_process() is an S3 that either calls das_read() on a
    character vector, or das_read() data frame as input and
    outputs a data frame with additional columns, e.g. for Mode or Bft
  TODO: describe columns created?
  TODO: check for events..?
  # TODO 1? If/how to incorporate these assumptions
  Effort assumptions; Specified by Jeff Moore June 2018 for CruzPlot for sake of early cruises
    \code{Mode[is.na(Mode)] <- "C"}
    \code{EffType[is.na(EffType)] <- "S"}

 TODO 2? Add warning() checks for unexpected values, e.g. for EffType?
   any(!(EffType %in% c("S", "N", "F", NA)))
   Ties into broader question of how much data checking should this do?

  Assumptions made during processing
  \itemize{
    \item Specific to ETP cruises at the moment, e.g. cruise.num.exp
    \item Determines effort using both '.' (OnEffort) and R and E events (OnEffort_RE)
    \item Resets all data at the beginning of each cruise (ID'd using days.gap)
    \item Mode letter is capitalized
    \item Glare: TRUE if HzSun = 11, 12 or 1 and VtSun = 2 or 3, or if HzSun = 12 and VtSun = 1; otherwise FALSE.
    \item If Glare is NA, i.e. if HorizSun or VertSun is NA, then Glare is FALSE
    \item RainFog is TRUE if 2, 3, or 4; otherwise FALSE
  }
  Note that das_process returns missing values as \code{NA} rather than \code{-1}
}
\examples{
# TODO
# x <- das_read("Data/1986-2006ETP.das")
# y <- das_process(x, reset.day = TRUE, days.gap = 50)
}
\seealso{
For more details about WinCruz and expected DAS data format, see
  \url{https://swfsc.noaa.gov/uploadedFiles/Divisions/PRD/WinCruz.pdf}
}
