% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_read.R
\name{das_read}
\alias{das_read}
\title{Read DAS file}
\usage{
das_read(file, tz = "UTC")
}
\arguments{
\item{file}{filename of a DAS file}

\item{tz}{character; see \link[base]{strptime}. Default is UTC}
}
\value{
Data frame with aerial DAS data parsed into columns.
  The following columns were processed or added:
  \itemize{
    \item \code{EffortDot}: logical; \code{TRUE} if "." present, \code{FALSE otherwise}
    \item \code{DateTime}: \code{POSIXct}; combination of \code{Date} and \code{Time} columns
    \item \code{Lat}: \code{numeric}; decimal degrees in range [-90, 90]
    \item \code{Lon}: \code{numeric}; decimal degrees in range [-180, 180]
    \item \code{Data#}: leading/trailing whitespace trimmed for non-comment events (rows where \code{Event} is not "C" or "c")
    \item \code{file_das}: character; filename (note not full filepath)
    \item \code{line_num}: integer; line number of each data row
  }
}
\description{
Read fixed-width DAS textfile generated by WinCruz
}
\details{
Parses DAS data into columns; adapted
  from \code{\link[swfscMisc]{das.read}}.
  The provided file must follow the following column number and format
  specifications (note that 'Data#' columns may be referred to as
  'Field#' columns in other documentation):
  \tabular{lrr}{
    \emph{Item} \tab \emph{Columns} \tab \emph{Format}\cr
    Sequence (event number) \tab 1-3\cr
    Event \tab 4\cr
    Effort dot \tab 5\cr
    Time \tab 6-11 \tab HHMMSS or HHMM\cr
    Date \tab 13-18 \tab MMDDYY\cr
    Latitude \tab 20-28 \tab NDD:MM.MM\cr
    Longitude \tab 30-39 \tab WDDD:MM.MM\cr
    Data1 \tab 40-44 \cr
    Data2 \tab 45-49 \cr
    Data3 \tab 50-54\cr
    Data4 \tab 55-59\cr
    Data5 \tab 60-64\cr
    Data6 \tab 65-69\cr
    Data7 \tab 70-74\cr
    Data8 \tab 75+\cr
  }

  TODO: handle non-UTF-8 symbols in comments.
}
\examples{
# TODO
# x <- das_read("../DAS_files/RV-Data/1986/MOPS0989.das")
# x <- das_read("../DAS_files/1986-2007ETP.das")

}
