% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_read.R
\name{das_read}
\alias{das_read}
\title{Read DAS file}
\usage{
das_read(file)
}
\arguments{
\item{file}{filename of a DAS file}
}
\value{
Data frame with columns for each DAS item...; idx_line only column added
  Processing done: create DateTime column and convert lat/lon to decimal;
  all else simply reading data from columns
  TODO: make return object of class das_read?
}
\description{
Read fixed-width DAS textfile generated by WinCruz
}
\details{
Read...
  Adapted from \code{\link[swfscMisc]{das.read}}
 TODO: describe columns created?
 TODO: check for events..?
}
\examples{
# TODO
# x <- das_read("Data/1986-2006ETP.das")

}
