% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_segdata.R
\name{das_segdata}
\alias{das_segdata}
\alias{das_segdata.data.frame}
\alias{das_segdata.das_df}
\title{Summarize DAS data for a continuous effort section}
\usage{
das_segdata(x, ...)

\method{das_segdata}{data.frame}(x, ...)

\method{das_segdata}{das_df}(x, conditions, segdata.method, seg.lengths, section.id, ...)
}
\arguments{
\item{x}{\code{das_df} object,
or a data frame that can be coerced to a \code{das_df} object.
Must contain a single continuous effort section of DAS data;
see the Details section below}

\item{...}{ignored}

\item{conditions}{see \code{\link{das_effort}}, or
see Details section for more information}

\item{segdata.method}{character; either \code{"avg"} or \code{"maxdist"}.
\code{"avg"} means the condition values will be
calculated as a weighted average by distance, while
\code{"maxdist"} means the condition values will be those recorded
for the longest distance during that segment}

\item{seg.lengths}{numeric; length of the modeling segments
into which \code{x} will be chopped}

\item{section.id}{numeric; the ID of \code{x} (the current continuous effort section)}
}
\value{
Data frame with the segdata information described above
  and in \code{\link{das_effort}}
}
\description{
Summarize DAS effort data by effort segment, while averaging conditions
}
\details{
This function was designed to be called by \code{\link{das_chop_equal}};
  users should avoid calling it themselves.
  It loops through the events in \code{x}, chopping \code{x} into modeling segments
  while calculating and storing relevant information for each segment.
  Because \code{x} is a continuous effort section, \code{x} must begin with
  a "B" or "R" event and end with the corresponding "E" event.

  For each segment, this function reports the segment number,
  segment ID, cruise number, the start/end/midpoints (lat/lon), segment length,
  year, month, day, time, mode, effort type,
  effective strip width sides (number of sides searched), and average conditions.
  The segment ID is designated as \code{eff_id} _ index of the modeling segment.
  Thus, if \code{section.id} is \code{1}, then the segment ID for
  the second segment from \code{x} is \code{"1_2"}.

  The average condition values are calculated as a weighted average by distance,
  and reported for the conditions specified by the \code{conditions} argument.
  For logical columns (e.g. Glare), the reported value is the percentage
  (in decimals) of the segment in which that condition was \code{TRUE}.

  Cruise number, mode, effort type, sides searched, and file name are
  also included in the segdata output.
  These values should all be consistent across the entire effort section,
  and thus across all segments in \code{x};
  a warning is printed if there are any inconsistencies

  \code{\link[swfscMisc]{bearing}} and \code{\link[swfscMisc]{destination}}
  are used to calculate the segment start, mid, and end points.
  TODO: option to pass distance method to swfscMisc functions?
  How does EAB/Eiren do these calculations?
}
\keyword{internal}
