% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_sight.R
\name{das_sight}
\alias{das_sight}
\alias{das_sight.data.frame}
\alias{das_sight.das_df}
\title{DAS sightings}
\usage{
das_sight(x, ...)

\method{das_sight}{data.frame}(x, ...)

\method{das_sight}{das_df}(x, returnformat = c("default", "wide", "comprehensive"), ...)
}
\arguments{
\item{x}{\code{das_df} object; output from \code{\link{das_process}},
or a data frame that can be coerced to a \code{das_df} object}

\item{...}{ignored}

\item{returnformat}{character; one of "default", "long", or "comprehensive"}
}
\value{
Data frame with 1) the columns from \code{x}, excluding the 'Data#' columns,
  and 2) columns with sighting information extracted from 'Data#' columns (described below).
  See \code{\link{das_format_pdf}} for more information the sighting information.
  If \code{returnformat} is "default", then there is one row for each species of each sighting;
  if \code{returnformat} is "wide", then there is one row for each sighting event;
  if \code{returnformat} is "complete", then there is one row for every
  group size estimate for each sighting.
  The specific columns in each output are described below.

  Sighting information columns in every format:
  \tabular{lll}{
    \emph{Sighting information}     \tab \emph{Column name} \tab \emph{Notes}\cr
    Sighting number                 \tab SightNo\cr
    Subgroup code                   \tab Subgroup\cr
    Observer that made the sighting \tab Obs\cr
    Standard observer               \tab Obs_std \tab Logical; \code{TRUE} if Obs is one of ObsL, Rec or ObsR,
      and \code{FALSE} if Obs is \code{NA}\cr
    Bearing to the sighting         \tab Bearing \tab Degrees, range 0 to 360\cr
    Number of reticle marks         \tab Reticle\cr
    Distance (nautical miles)       \tab DistNm\cr
    Sighting cue                    \tab Cue\cr
    Sighting method                 \tab Method\cr
    Photos of school?               \tab Photos\cr
    Birds present with school?      \tab Birds\cr
    Calibration school?             \tab CalibSchool\cr
    Aerial photos taken?            \tab PhotosAerial\cr
    Biopsy taken?                   \tab Biopsy\cr
    Probable sighting               \tab Prob  \tab Logical; \code{NA} for non-S/K/M/G events\cr
    Number of species in sighting   \tab nSp   \tab \code{NA} for non-S/K/M/G events\cr
    Mixed species sighting          \tab Mixed \tab Logical; \code{TRUE} if nSp > 1\cr
    Course (true heading) of school at resight \tab CourseSchool \tab \code{NA} for non-s/k/m events\cr
    Presence of associated JFR  \tab TurtleJFR  \tab \code{NA} for non-"t" events; JFR = jellyfish, floating debris, or red tide \cr
    Estimated turtle maturity   \tab TurtleAge  \tab \code{NA} for non-"t" events\cr
    Perpendicular distance (km) to sighting \tab PerpDistKm \tab Calculated via \code{(abs(sin(Bearing*pi/180) * DistNm) * 1.852)}\cr
  }
  To convert the perpendicular distance back to nautical miles,
  one would divide PerpDistKm by 1.852.

  Sighting information columns present specifically in the "default" format output:
  \tabular{lll}{
    School group size - best estimate \tab GsSchoolBest \tab Arithemtic mean of all best estimates\cr
    School group size - high estimate \tab GsSchoolHigh \tab Arithmetic mean of all high estimates\cr
    School group size - low estimate  \tab GsSchoolLow  \tab Arithmetic mean of all low estimates\cr
    Species code \tab SpCode  \tab For mammal, turtle, and pinniped species codes, as well as boat type\cr
    Species-specific group size - best estimate \tab GsSpBest \tab
      GsSchoolBest multiplied by the (average) corresponding species percentage \cr
  }

  Sighting information columns present specifically in the "wide" format output:
  \tabular{lll}{
    Species 1 code                \tab SpCode1\cr
    Species 2 code                \tab SpCode2\cr
    Species 3 code                \tab SpCode3\cr
    Species 4 code                \tab SpCode4\cr
    Percentage of Sp1 in sighting \tab SpPerc1\cr
    Percentage of Sp2 in sighting \tab SpPerc2\cr
    Percentage of Sp3 in sighting \tab SpPerc3\cr
    Percentage of Sp4 in sighting \tab SpPerc4\cr
    Species 1 group size          \tab GsSpBest1\cr
    Species 2 group size          \tab GsSpBest2\cr
    Species 3 group size          \tab GsSpBest3\cr
    Species 4 group size          \tab GsSpBest4\cr
    Species 1 probable code       \tab SpProb1 \tab Extracted from '?' event\cr
    Species 2 probable code       \tab SpProb2 \tab Extracted from '?' event\cr
    Species 3 probable code       \tab SpProb3 \tab Extracted from '?' event\cr
    Species 4 probable code       \tab SpProb4 \tab Extracted from '?' event\cr
    Turtle species             \tab TurtleSp    \tab \code{NA} for non-"t" events\cr
    Number of turtles          \tab TurtleNum   \tab \code{NA} for non-"t" events\cr
    Was turtle captured?       \tab TurtleCapt  \tab \code{NA} for non-"t" events\cr
    Pinniped species           \tab PinnipedSp  \tab \code{NA} for non-"p" events\cr
    Number of pinnipeds        \tab PinnipedNum \tab \code{NA} for non-"p" events\cr
    Boat or gear type          \tab BoatType    \tab \code{NA} for non-"F" events\cr
    Number of boats            \tab BoatNum     \tab \code{NA} for non-"F" events\cr
  }

  Sighting information columns present specifically in the "complete" format output:
  \tabular{lll}{
    TODO
  }
}
\description{
Extract sightings and associated information from processed DAS data
}
\details{
DAS events contain specific information in the 'Data#' columns,
  with the information depending on the event code for that row.
  The output data frame contains columns with this specific information
  extracted to dedicated columns as described below.
  This function recognizes the following types of sightings:
  marine mammal sightings (event codes "S", "K", or "M"),
  marine mammal resights (codes "s", "k", "m"),
  marine mammal subgroup sightings (code "G"),
  marine mammal subgroup resights (code "g"),
  turtle sightings (code "t"),
  pinniped sightings (code "p")
  and fishing vessel sightings (code "F").
  Warnings are printed if all S, K, M, and G events (and only these events) are followed by
  an A event and at least one numeric event.
  See \code{\link{das_format_pdf}} for more information about events and event formats.

  \code{returnformat} "default": One row per species; multi-species sighitngs are split into multiple lines.
    Turtle num/sp and Boat num combined into Sp and Gs columns
  \code{returnformat} "wide": One column for everything, one row for each sighting event. Values averaged for different observers
  \code{returnformat} "comprehensive": TODO

  Abbreviations used in column names: Gs = group size, Sp = species,
  Nm = nautical mile, Perc = percentage, Prob = probable

  This function makes the following assumptions, and alterations to the raw data:
  \itemize{
    \item "G", "S", "K", and "M" events, and only these events,
      are immediately followed by an "A" event.
    \item "A" events following an S/K/M/G event have the same Data1 (sighting number)
      as the S/K/M/G event.
    \item The 'nSp' column is equivalent to the number of non-\code{NA}
      'Data5', 'Data6', 'Data7', and 'Data8' values
      for the corresponding "A" event are not \code{NA}, and \code{FALSE} otherwise.
      This column has a non-\code{NA} value for only "S", "K", "M", and "G" events
    \item The 'Prob' column is \code{TRUE} is the sighting has an associated "?" event,
      and \code{FALSE} otherwise.
      This column has a non-\code{NA} value for only "S", "K", and "M" events
    \item The 'GsSchoolBest', 'GsSchoolHigh', and 'GsSchoolLow' columns are the arithmetic mean of
      the 'Data2', 'Data3', and 'Data4' columns, respectively, with \code{na.rm = TRUE}
      for the associated "1"-"8" events
    \item The 'SpPerc1', 'SpPerc2', 'SpPerc3', and 'SpPerc4' values are the
      arithemtic means of the 'Data5', 'Data6', 'Data7', and 'Data8' columns (with \code{na.rm = TRUE}),
      respectively, for the associated "1"-"8" events
    \item The 'GsSpBest#' values are the product of the 'GsSchoolBest' and the respective 'SpPerc#' columns
    \item The values for the following columns are capitalized using
      \code{\link[base:chartr]{toupper}}:
      'Birds', 'Photos', 'CalibSchool', 'PhotosAerial', 'Biopsy',
      'TurtleAge', and 'TurtleCapt'
  }
}
\examples{
y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)

das_sight(y.proc)
das_sight(y.proc, returnformat = "wide")

}
