% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_sight.R
\name{das_sight}
\alias{das_sight}
\title{DAS sightings}
\usage{
das_sight(das.df, mixed.multi)
}
\arguments{
\item{das.df}{data frame; processed DAS data.. TODO}

\item{mixed.multi}{logical; indicates if mixed-species sightings should be output in multiple rows}
}
\value{
Data frame with ...
  Associated information (observer, species, etc.) is extracted
  from \code{Data#} columns as specified in Details.
  Other columns from \code{das.df} will be included in the output
}
\description{
Extract sightings and associated information from processed DAS data
}
\details{
This function requires that the following assumptions be true:

  \itemize{
    \item All (and only) "S", "K", and "M" events are immediately followed by an "A" event
    \item TODO
  }
}
\examples{
#TODO

}
