% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_sight.R
\name{das_sight}
\alias{das_sight}
\alias{das_sight.data.frame}
\alias{das_sight.das_df}
\title{DAS sightings}
\usage{
das_sight(x, mixed.multi)

\method{das_sight}{data.frame}(x, mixed.multi)

\method{das_sight}{das_df}(x, mixed.multi)
}
\arguments{
\item{x}{\code{das_df} object; output from \code{\link{das_process}}}

\item{mixed.multi}{logical; indicates if mixed-species sightings should be output in multiple rows}
}
\value{
Data frame with ...
  Associated information (observer, species, etc.) is extracted
  from \code{Data#} columns as specified in Details.
  Other columns from \code{x} will be included in the output
}
\description{
Extract sightings and associated information from processed DAS data
}
\details{
This function requires that the following assumptions be true:

  \itemize{
    \item All (and only) "S", "K", and "M" events are immediately followed by an "A" event
    \item TODO
  }
}
\examples{
y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)

das_sight(y.proc)

}
