% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_sight.R
\name{das_sight}
\alias{das_sight}
\alias{das_sight.data.frame}
\alias{das_sight.das_df}
\title{DAS sightings}
\usage{
das_sight(x, ...)

\method{das_sight}{data.frame}(x, ...)

\method{das_sight}{das_df}(x, returnformat = c("default", "wide", "comprehensive"), ...)
}
\arguments{
\item{x}{\code{das_df} object; output from \code{\link{das_process}},
or a data frame that can be coerced to a \code{das_df} object}

\item{...}{ignored}

\item{returnformat}{character; one of "default", "long", or "comprehensive"}
}
\value{
Data frame with 1) the columns from \code{x}, excluding the 'Data#' columns,
  and 2) columns with sighting information extracted from 'Data#' columns (described below).
  See \code{\link{das_format_pdf}} for more information the sighting information.
  The data frame has one row for each sighting,
  or one row for each species of each sighting if \code{mixed.multi} is \code{TRUE}.

  Added sighting information columns:
  \tabular{lll}{
    \emph{Sighting information}     \tab \emph{Column name} \tab \emph{Notes}\cr
    Sighting number                 \tab SightNo\cr
    Subgroup code                   \tab Subgroup\cr
    Observer that made the sighting \tab Obs\cr
    Standard observer               \tab Obs_std \tab Logical; \code{TRUE} if Obs is one of ObsL, Rec or ObsR,
      and \code{FALSE} if Obs is \code{NA}\cr
    Bearing to the sighting         \tab Bearing \tab Degrees, range 0 to 360\cr
    Number of reticle marks         \tab Reticle\cr
    Distance (nautical miles)       \tab DistNm\cr
    Sighting cue                    \tab Cue\cr
    Sighting method                 \tab Method\cr
    Photos of school?               \tab Photos\cr
    Birds present with school?      \tab Birds\cr
    Probable sighting               \tab Prob  \tab Logical\cr
    Number of species in sighting   \tab nSp   \tab \code{NA} for non-S/K/M/G events\cr
    Mixed species sighting          \tab Mixed \tab Logical; \code{TRUE} if nSp > 1\cr
    Total group size                \tab GsTotal \tab Only different from GsSp if mixed species sighting\cr
    Species  code                 \tab Sp      \tab Only present when \code{mixed.multi = TRUE}\cr
    Species-specific group size   \tab GsSp    \tab Only present when \code{mixed.multi = TRUE}\cr
    Species 1 code                \tab Sp1     \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 2 code                \tab Sp2     \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 3 code                \tab Sp3     \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 4 code                \tab Sp4     \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 1-specific group size \tab Sp1Perc \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 2-specific group size \tab Sp2Perc \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 3-specific group size \tab Sp3Perc \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 4-specific group size \tab Sp4Perc \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 1 group size          \tab GsSp1   \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 2 group size          \tab GsSp2   \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 3 group size          \tab GsSp3   \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 4 group size          \tab GsSp4   \tab Only present when \code{mixed.multi = FALSE}\cr
    Species 1 probable code       \tab ProbSp1 \tab From '?' event, only present when \code{mixed.multi = FALSE}\cr
    Species 2 probable code       \tab ProbSp2 \tab From '?' event, only present when \code{mixed.multi = FALSE}\cr
    Species 3 probable code       \tab ProbSp3 \tab From '?' event, only present when \code{mixed.multi = FALSE}\cr
    Species 4 probable code       \tab ProbSp4 \tab From '?' event, only present when \code{mixed.multi = FALSE}\cr
    Course of resight group    \tab ResightCourse \tab \code{NA} for non-resight events\cr
    Turtle species             \tab TurtleSp    \tab \code{NA} for non-"t" events\cr
    Number of turtles          \tab TurtleNum   \tab \code{NA} for non-"t" events\cr
    Presence of associated JFR \tab TurtleJFR   \tab \code{NA} for non-"t" events; JFR = jellyfish, floating debris, or red tide \cr
    Estimated turtle maturity  \tab TurtleAge   \tab \code{NA} for non-"t" events\cr
    Was turtle captured?       \tab TurtleCapt  \tab \code{NA} for non-"t" events\cr
    Pinniped species           \tab PinnipedSp  \tab \code{NA} for non-"p" events\cr
    Number of pinnipeds        \tab PinnipedNum \tab \code{NA} for non-"p" events\cr
    Boat or gear type          \tab BoatType    \tab \code{NA} for non-"F" events\cr
    Number of boats            \tab BoatNum     \tab \code{NA} for non-"F" events\cr
    Perpendicular distance (km) \tab PerpDistKm \tab Calculated via \code{(abs(sin(Bearing*pi/180) * DistNm) * 1.852)}\cr
  }

  To convert the perpendicular distance back to nautical miles,
  one would divide PerpDistKm by 1.852.
}
\description{
Extract sightings and associated information from processed DAS data
}
\details{
DAS events contain specific information in the 'Data#' columns,
  with the information depending on the event code for that row.
  The output data frame contains columns with this specific information
  extracted to dedicated columns as described below.
  This function recognizes the following types of sightings:
  marine mammal sightings (event codes "S", "K", or "M"),
  marine mammal resights (codes "s" or "k"),
  marine mammal subgroup sightings (code "G"),
  marine mammal subgroup resights (code "g"),
  turtle sightings (code "t"),
  pinniped sightings (code "p")
  and fishing vessel sightings (code "F").
  See \code{\link{das_format_pdf}} for more information about events and event formats.

  \code{returnformat} "default": One row per species; multi-species sighitngs are split into multiple lines.
    Turtle num/sp and Boat num combined into Sp and Gs columns
  \code{returnformat} "wide": One column for everything, one row for each sighting event. Values averaged for different observers
  \code{returnformat} "comprehensive": TODO


  Abbreviations used in column names: Gs = group size, Sp = species,
  Nm = nautical mile, Perc = percentage, Prob = probable

  This function makes the following assumptions, and alterations to the raw data:
  \itemize{
    \item "G", "S", "K", and "M" events, and only these events,
      are immediately followed by an "A" event.
    \item "A" events following an S/K/M event have the same Data1 (sighting number)
      as the S/K/M event.
      "A" events following a "G" event have a character (e.g. "A" or "B")
    \item The 'nSp' column is equivalent to the number of non-\code{NA}
      'Data5', 'Data6', 'Data7', and 'Data8' values
      for the corresponding "A" event are not \code{NA}, and \code{FALSE} otherwise.
      This column has a non-\code{NA} value for only "S", "K", and "M" events
    \item The 'Prob' column is \code{TRUE} is the sighting has an associated "?" event,
      and \code{FALSE} otherwise.
      This column has a non-\code{NA} value for only "S", "K", and "M" events
    \item The 'GsTotal' column is the mean of the 'Data2' columns (with \code{NA}s removed)
      for the associated "1"-"8" events
    \item The 'Sp1Perc', 'Sp2Perc', 'Sp3Perc', and 'Sp4Perc' values are the
      means of the 'Data5', 'Data6', 'Data7', and 'Data8' columns (with \code{na.rm = TRUE}),
      respectively, for the associated "1"-"8" events
    \item The 'GsSp1', 'GsSp2', 'GsSp3', and 'GsSp4' values are the product of
      the 'GsTotal' and the respective '...Perc' columns
    \item The values for the following columns were capitalized using
      \code{\link[base:chartr]{toupper}}:
      'Birds', 'Photos', 'TurtleAge', and 'TurtleCapt'
  }

  Outstanding questions/todo:
  \itemize{
    \item Should any columns be converted to logicals, e.g. 'Birds', 'Photos', and 'TurtleCapt'?
    \item Should \code{NA} values for columns 'Sp1', 'Sp1Perc', 'GsSp1', etc.,
      be changed from \code{NA} to \code{0}?
    \item TODO: Add flag for option to have comprehensive output of group size estimations
  }
}
\examples{
y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)

das_sight(y.proc)
das_sight(y.proc, returnformat = "wide")

}
