% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randpicks_convert.R
\name{randpicks_convert}
\alias{randpicks_convert}
\title{Convert randpicks file}
\usage{
randpicks_convert(x.randpicks, x.segdata, seg.km)
}
\arguments{
\item{x.randpicks}{Data frame with two columns;
randpick values formatted for segchopr that correspond to \code{x.segdata}}

\item{x.segdata}{Data frame; segdata that corresponds to \code{x.randpicks}}

\item{seg.km}{numeric; target segment length used when creating \code{x.segdata}}
}
\value{
Data frame with one line for each continuous effort section,
  and two columns: \code{effort_section} and \code{randpicks}
}
\description{
Convert randpicks file from segchopr format to swfscAirDAS format
}
\details{
Past DAS processing code used a different randpicks format,
  where only the generated random values were recorded.
  This function 'converts' a data frame with those values to a data frame
  in the format required by this package
}
\examples{
#TODO
#randpicks_convert(eab.randpicks, eab.segdata, 5)

}
