% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swfscDAS-package.R
\docType{package}
\name{swfscDAS-package}
\alias{swfscDAS-package}
\alias{swfscDAS}
\title{Southwest Fisheries Science Center DAS}
\description{
Process and summarize shipboard DAS data
}
\details{
This package contains functions designed for processing and analyzing DAS data
generated using the WinCruz program by the Southwest Fisheries Science Center.
It is intended to standardize and streamline basic DAS data processing.
}
\seealso{
\url{https://swfsc.github.io/swfscDAS/}
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{package}
