% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swfscDAS-package.R
\docType{package}
\name{swfscDAS-package}
\alias{swfscDAS-package}
\alias{swfscDAS}
\title{Southwest Fisheries Science Center DAS}
\description{
Process and summarize shipboard DAS data
}
\details{
This package contains functions designed for processing and analyzing DAS data
generated using the WinCruz program by the Southwest Fisheries Science Center.
Functionality currently includes reading DAS data into a data frame,
processing this data (extracting state and condition information for each DAS event),
and summarizing sighting and effort information.
}
\seealso{
\url{https://smwoodman.github.io/swfscDAS/}
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{package}
